requirements.txt
# ============================================================
# TriageIA — Dependencias del proyecto (ML/NLP/análisis)
# Instalar: pip install -r requirements.txt
# Modelo spaCy: python -m spacy download en_core_web_sm
#
# NOTA: Las dependencias de Airflow están separadas en:
#       infra/airflow/requirements-airflow.txt
#       para no contaminar el entorno de notebooks/ML.
# ============================================================

# --- Core ML ---
scikit-learn>=1.3.0
xgboost>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
imbalanced-learn>=0.11.0  # SMOTE y técnicas para desbalanceo

# --- NLP ---
spacy>=3.7.0
nltk>=3.8.0

# --- Visualización ---
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# --- Explicabilidad ---
shap>=0.43.0

# --- Audio a texto (demo Streamlit) ---
openai-whisper>=20231117

# --- App demo ---
streamlit>=1.28.0

# --- LLM (extracción/normalización en Phase 3-4, NO clasificación final) ---
anthropic>=0.25.0

# --- Base de datos (trazabilidad en Postgres) ---
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# --- Almacenamiento de objetos (MinIO/S3) ---
minio>=7.2.0

# --- Notebooks ---
jupyterlab>=4.0.0
ipykernel>=6.25.0

# --- Utilidades ---
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.65.0
joblib>=1.3.0
